refactor(presentation): extract obstacles + walls from PresentationSpawnSystem (3576 -> 2991)
Fix #204 (first pass).

PresentationSpawnSystem.cs was 3576 lines with procedural generators for
forests, rocks, iron deposits, wall hubs, wall segments, wall instances,
wall towers, wall gates, smelters, halls, huts, barracks, vaults, cursed
ground, cadaver loot, crystal entities, crystal nodes — plus transform
syncing, cleanup, faction color logic, and banner placement.

Marked the class as `partial` and extracted the two largest, most self-
contained sections into sibling files:

  PresentationSpawnSystem.Obstacles.cs (286 lines)
    Procedural forest generation (patches, trunks, canopies) and
    procedural rock cluster + iron deposit generation.

  PresentationSpawnSystem.Walls.cs (329 lines)
    Alanthor wall hub, wall segment, wall instance, wall tower, and
    wall gate procedural generation.

Main file is now 2991 lines (down from 3576, -16%). The remaining
sections (smelter, hall, hut, barracks, vault, crystal, cadaver, faction
color, transform sync) can be extracted incrementally using the same
partial-class pattern.

All three files brace-balanced. No behavioral changes — methods moved
verbatim with only the class wrapper and using directives added.

Closes #204
